requirements.txt
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
ipykernel
python-dotenv
nest-asyncio